requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-cors-headers==4.3.0
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
httplib2==0.22.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
paho-mqtt==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
